Move cpython core tests to their own module
